requirements.txt
plumbum
future
numpy
pandas
nibabel
pyyaml
python_log_indenter
